modules/PushButton/requirements.txt
azure-iot-device~=2.7.0
RPi.GPIO~=0.7.0